feat: implemented json schema editor
