+++ webapp/src/lib/components/iconButton.svelte
@@ -5,21 +5,28 @@
 
 	export let icon: IconComponent = XMark;
 	export let border = false;
+	export let size: keyof typeof sizes = 'xs';
+	export let href: string | undefined = undefined;
 
 	const sizes = {
 		xs: {
 			iconSize: 20,
 			padding: '!p-1'
+		},
+		sm: {
+			iconSize: 20,
+			padding: '!p-2'
 		}
 	};
 
-	const props = sizes['xs'];
+	$: props = sizes[size];
 </script>
 
 <Button
 	class={`${props.padding} ${!border ? '!border-transparent' : ''}`}
 	color="alternative"
 	on:click
+	{href}
 >
 	<svelte:component this={icon} size={props.iconSize.toString()} />
 </Button>
